requirements.txt
aioretry==5.0.2
async_modbus==0.2.1
pydantic==2.5.3
pyserial==3.5
PyYAML==6.0.1
PyYAML==6.0.1
sanic==23.12.1
sanic_ext==23.12.0
